requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
sqlalchemy==2.0.28
psycopg2-binary==2.9.9
alembic==1.13.1
pydantic==2.5.3
python-multipart==0.0.6
datetime
typing-extensions
